refactor: applied MVT architecture
logout, home, dashboard passed to views.py, created new oauth.py
