+++ backend/transcendence/urls.py
@@ -19,14 +19,14 @@ from django.urls import path
 from django.conf import settings
 from django.conf.urls.static import static
 
-from user_auth import requestAuth, logout, home, dashboard
+from user_auth import oauth, views
 
 
 urlpatterns = [
-    path('', home.home, name='home'),
+    path('', views.home, name='home'),
     path('admin/', admin.site.urls),
-    path('auth/42/', requestAuth.requestAuth42, name='requestAuth42'),
-    path('auth/42/callback/', requestAuth.callbackAuth, name='callbackAuth'),
-    path('logout/', logout.logout, name='logout'),
-    path('dashboard/', dashboard.dashboard, name='dashboard')
+    path('auth/42/', oauth.oauth42, name='oauth42'),
+    path('auth/42/callback/', oauth.callbackAuth, name='callbackAuth'),
+    path('logout/', views.logout, name='logout'),
+    path('dashboard/', views.dashboard, name='dashboard')
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
