--- backend/user_auth/home.py
@@ -1,6 +1,0 @@
-from django.shortcuts import redirect, render   #to use render() and requests()
-
-def home(request):
-    if request.user.is_authenticated:
-        return redirect('dashboard')
-    return render(request, 'home.html')
