+++ backend/user_auth/models.py
@@ -1,12 +1,8 @@
 from django.db import models
-from django.contrib.auth.models import AbstractUser
-from django.utils.timezone import now
 
 # Create your models here.
 from django.contrib.auth.models import AbstractBaseUser
 from django.contrib.auth.base_user import BaseUserManager
-from django.db import models
-
 from django.utils import timezone
 
 class CustomUserManager(BaseUserManager):
